fix(opencog): use size_t literals for std::min on Windows MSYS2
On Linux x86_64 (LP64) size_t and unsigned long are the same type, so
std::min(size_t, 4096UL) deduces fine. On MSYS2/UCRT64 (LLP64) size_t
is unsigned long long while UL literals are unsigned long, breaking
template type deduction:

  no matching function for call to 'min(const size_t&, long unsigned int)'

Cast the literal explicitly to size_t so both arguments share a type.

+++ src/opencog/llm_inference.cpp
@@ -54,7 +54,7 @@ bool LLMInferenceEngine::load_model(const std::string& model_path, const Archite
 
     // Set up context parameters
     llama_context_params ctx_params = llama_context_default_params();
-    ctx_params.n_ctx = std::min(config.max_context_length, 4096UL);
+    ctx_params.n_ctx = std::min(config.max_context_length, static_cast<size_t>(4096));
     ctx_params.n_threads = std::thread::hardware_concurrency();
     ctx_params.n_threads_batch = ctx_params.n_threads;
 
@@ -239,7 +239,7 @@ std::string LLMInferenceEngine::cognitive_inference(const std::string& input,
     cognitive_prompt << "Please provide a thoughtful response considering the knowledge context and goals:";
 
     // Generate response with architecture awareness
-    size_t max_tokens = std::min(config.max_context_length / 4, 512UL);
+    size_t max_tokens = std::min(config.max_context_length / 4, static_cast<size_t>(512));
     return generate_response(cognitive_prompt.str(), state.current_focus, max_tokens);
 }
